#64 rename: pot의 subdomain 생성
