--- src/main/java/org/com/itpple/spot/server/domain/pot/category/repository/CategoryRepository.java
@@ -1,8 +1,0 @@
-package org.com.itpple.spot.server.domain.pot.category.repository;
-
-import org.com.itpple.spot.server.domain.pot.category.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-
-}
