+++ src/main/java/org/com/itpple/spot/server/global/exception/code/ErrorCode.java
@@ -33,6 +33,7 @@ public enum ErrorCode {
     // Pot 관련
     NOT_FOUND_POT(1701, "팟을 찾을 수 없습니다.", HttpStatus.NOT_FOUND),
     NOT_FOUND_CATEGORY(1702, "카테고리를 찾을 수 없습니다.", HttpStatus.NOT_FOUND),
+    CONFLICT_HASHTAG(1703, "이미 존재하는 해시태그입니다.", HttpStatus.CONFLICT),
 
     // Reaction 관련
     NOT_ADD_MULTIPLE_REACTION(1801, "팟에 여러 개의 반응을 추가할 수 없습니다.", HttpStatus.BAD_REQUEST),
