+++ src/test/java/org/com/itpple/spot/server/domain/pot/repository/PotRepositoryTest.java
@@ -7,7 +7,7 @@ import static org.com.itpple.spot.server.global.util.GeometryUtil.createPolygon;
 
 import lombok.RequiredArgsConstructor;
 import org.com.itpple.spot.server.domain.location.dto.PointDTO;
-import org.com.itpple.spot.server.domain.pot.category.entity.Category;
+import org.com.itpple.spot.server.domain.pot.domain.category.entity.Category;
 import org.com.itpple.spot.server.domain.pot.entity.Pot;
 import org.com.itpple.spot.server.domain.user.entity.User;
 import org.com.itpple.spot.server.global.common.constant.PotType;
