+++ src/test/java/org/com/itpple/spot/server/util/PotTestUtil.java
@@ -1,8 +1,7 @@
 package org.com.itpple.spot.server.util;
 
 import java.time.LocalDateTime;
-
-import org.com.itpple.spot.server.domain.pot.category.entity.Category;
+import org.com.itpple.spot.server.domain.pot.domain.category.entity.Category;
 import org.com.itpple.spot.server.domain.pot.entity.Pot;
 import org.com.itpple.spot.server.domain.user.entity.User;
 import org.com.itpple.spot.server.global.common.constant.PotType;
